test: Move code for filtering metadata out of ui.tsx to facilitate unit testing
